Remove duplicate code for 32 bit code path
Use template and parameter to reduce code duplication.
The differences between the "upload" routines from amgx_c.cu were the type cast to
either int64_t or int for column indices.
This one out of two overloads of loadDistributedMatrix. The overloads only differ in choice of column
type, not in logic, so made column type a template parameter.
Only wrinkle is that another explicit instantiation is required (bottom of distributed_manager.cu), but
that seems in line with the general design.